usb: musb: Add a BUG_ON if trying to suspend musb in B_IDLE state
Musb sometimes failes to be suspended because it is active and in the
B_IDLE state. When this happens, USBPHY is left on and we don't hit OFF
mode. Add a BUG_ON to reboot in that case as it is better than emptying
the battery and so that we can see how often this happens.

Change-Id: Ic8d8c6918f3b9a8f0671f484d5d72fae63d7a61d
Signed-off-by: Benoit Goby <benoit@android.com>
(cherry picked from commit 8ebb62bd38adcbef29a457a064f61c1d9bb5526c)