OMAP4+: HSI: Prevent HSI clock disable while ACWAKE is high
If HSI clocks are disabled while ACWAKE is high, it is possible that ACWAKE
goes low unexpectedly.
This is not seen on Link Layers which do not comply with MIPI, because in this
case ACWAKE and CAWAKE are high at the same time, so HSI clocks are not
disabled.
But for Link Layers which comply with MIPI, CAWAKE will go low independtly of
CAWAKE, and HSI driver might cut the clocks while ACWAKE is high, which causes
CAWAKE to have an unknwon value.

Bug: 6217861, 6207265
Change-Id: I3098619a645277af12eb0ba82ca0e1508bd3ce03
Signed-off-by: Djamil Elaidi <d-elaidi@ti.com>
Signed-off-by: wonhee <wonhee48.seo@samsung.com>
(cherry picked from commit 52c33735e828ee7fb6e8f0393a048876788209b2)