Enable gamma feature in DSS
This patch enables the Gamma table to be loaded. sysfs entries are used to
control the value of gamma ranging from 0.2 to 2.0 in steps of 0.2.
The different values of gamma can also be used to alter the brightness
level of image.

Change-Id: I072bc5465ad636bfb1a416db1534542dfd5fa421
Signed-off-by: Sujeet Barnawal <s-baranwal@ti.com>
(cherry picked from commit 328c61c426dfa277b1a731a5bab50f499945ff45)